moved working files to ChapterRepresentation class; refactored Narration load
