+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/domain/narration/NarrationHistory.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.otter.common.data.narration
+package org.wycliffeassociates.otter.common.domain.narration
 
 import org.wycliffeassociates.otter.common.audio.AudioFile
 import org.wycliffeassociates.otter.common.data.primitives.VerseNode
